Refactor server code
The code that added an immutable cache header to the static files didn't
actually do anything. So removing it to simplify things
