+++ packages/admin-frontend/server/index.js
@@ -1,18 +1,47 @@
 // @flow
 /**
  * Part of GDL gdl-frontend.
- * Copyright (C) 2018 GDL
+ * Copyright (C) 2017 GDL
  *
  * See LICENSE
  */
+
+const express = require('express');
+const next = require('next');
+const cookieParser = require('cookie-parser');
+const { GDL_ENVIRONMENT } = require('gdl-config');
 
 const {
   serverRuntimeConfig: { port }
 } = require('../config');
 
-const initServer = require('./server');
+const isDev = process.env.NODE_ENV !== 'production';
+const app = next({ dev: isDev });
 
-initServer().then(server => {
+app.prepare().then(() => {
+  const server = express();
+
+  if (isDev || GDL_ENVIRONMENT === 'dev') {
+    app.setAssetPrefix(`http://localhost:${port}`);
+  } else if (GDL_ENVIRONMENT === 'local') {
+    // NB! If you try to access admin-frontend through the nginx proxy locally
+    // you probably want to comment out this part and set /admin as the prefix
+    app.setAssetPrefix('http://localhost:40006/');
+  } else {
+    app.setAssetPrefix('/admin');
+  }
+
+  // Health check for AWS
+  // $FlowFixMe: https://github.com/flowtype/flow-typed/issues/1120
+  server.get('/health', (req, res) => {
+    res.status(200).json({ status: 200, text: 'Health check ok' });
+  });
+
+  // Setup the cookie parsing for express
+  server.use(cookieParser());
+
+  server.use(app.getRequestHandler());
+
   server.listen(port, err => {
     if (err) throw err;
     console.log(`> Ready on http://localhost:${port}`);
